MAJOR UPDATE (Warning this could break things):  Migrated the code to a Python 3 compatible form.  Note that this should still work on Python 2.7
Most changes were made using 2to3, then the imports had to be fixed manually.

Finally, since Jupyter as now replaced IPython, the code was adapted to use Jupyter's latest API.  This also fixes some of the awkwardness of the starting a kernel and connecting the app to it via an Console.  The Data_Analyzer App as not yet been modify to reflect those changes.